ControllerCharacter
물리를 사용했던 기능은 이전과 동일하나 로직을 수정해봄 코드가 많이 단축됨
물리를 사용하지 않기에 계속해서 물리 계산을 수행하는 리지드바디보다 퍼포먼스도 더욱 효율적
