+++ Assets/Scripts/Character/ControllerCharacter.cs.meta
